Add example and docs
