+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/Iterate/Visibility.mdx
@@ -1,0 +1,27 @@
+---
+title: 'Visibility'
+description: 'The `Iterate.Visibility` component allows you to conditionally display content based on relative paths (`itemPath`) within an `Iterate.Array` component.'
+showTabs: true
+tabs:
+  - title: Info
+    key: '/info'
+  - title: Demos
+    key: '/demos'
+  - title: Properties
+    key: '/properties'
+  - title: Events
+    key: '/events'
+breadcrumb:
+  - text: Forms
+    href: /uilib/extensions/forms/
+  - text: Iterate
+    href: /uilib/extensions/forms/Iterate/
+  - text: Iterate.Visibility
+    href: /uilib/extensions/forms/Iterate/Visibility/
+---
+
+import Info from 'Docs/uilib/extensions/forms/Iterate/Visibility/info'
+import Demos from 'Docs/uilib/extensions/forms/Iterate/Visibility/demos'
+
+<Info />
+<Demos />
