+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/Iterate/Visibility/demos.mdx
@@ -1,0 +1,11 @@
+---
+showTabs: true
+---
+
+import * as Examples from './Examples'
+
+## Demos
+
+### Basic example
+
+<Examples.BasicExample />
