+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/Iterate/Visibility/events.mdx
@@ -1,0 +1,10 @@
+---
+showTabs: true
+---
+
+import PropertiesTable from 'dnb-design-system-portal/src/shared/parts/PropertiesTable'
+import { VisibilityEvents } from '@dnb/eufemia/src/extensions/forms/Form/Visibility/VisibilityDocs'
+
+## Events
+
+<PropertiesTable props={VisibilityEvents} />
