+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/Iterate/Visibility/info.mdx
@@ -1,0 +1,9 @@
+---
+showTabs: true
+---
+
+## Description
+
+The `Iterate.Visibility` component allows you to conditionally display content based on relative paths (`itemPath`) within an [Iterate.Array](/uilib/extensions/forms/Iterate/Array/) component.
+
+Fore more details, head over to the [Form.Visibility](/uilib/extensions/forms/Form/Visibility/) component documentation.
