+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/Iterate/Visibility/properties.mdx
@@ -1,0 +1,10 @@
+---
+showTabs: true
+---
+
+import PropertiesTable from 'dnb-design-system-portal/src/shared/parts/PropertiesTable'
+import { VisibilityProperties } from '@dnb/eufemia/src/extensions/forms/Form/Visibility/VisibilityDocs'
+
+## Properties
+
+<PropertiesTable props={VisibilityProperties} />
