+++ packages/dnb-eufemia/src/extensions/forms/Iterate/Visibility/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Visibility'
+export * from './Visibility'
