+++ packages/dnb-eufemia/src/extensions/forms/Iterate/index.ts
@@ -9,6 +9,7 @@ export { default as EditContainer } from './EditContainer'
 export { default as ViewContainer } from './ViewContainer'
 export { default as AnimatedContainer } from './AnimatedContainer'
 export { default as Toolbar } from './Toolbar'
+export { default as Visibility } from './Visibility'
 export { default as ItemNo } from './ItemNo'
 export { useCount, count, Count } from './Count'
 export { default as useItem } from './hooks/useItem'
